Refactor Agent; use Pair OD, Float distances
Major refactor of cityimage Agent: new constructor takes a RouteChoice and a List<Pair<NodeGraph,NodeGraph>> OD, stores routeChoice, adds setOD and placeAtOriginWithoutLayerUpdate (uses GeometryFactory), simplifies step/handleReachedDestination logic (increments trips, stops/removes agent appropriately), updates route planning/logging, and adds getAgentScenario/getRouteChoice accessors. Also changed PedSimCityImage.distances from Double to Float and updated Import to parse floats. Populate now selects start nodes via NodesLookup.randomNodeFromGeometries(...) instead of indexing PedSimCity.startingNodes.

+++ src/main/java/pedsim/cityimage/engine/Import.java
@@ -62,7 +62,7 @@ public class Import extends pedsim.core.engine.Import {
 			row += 1;
 			if (row == 1)
 				continue; // Skip header
-			PedSimCityImage.distances.add(Double.parseDouble(nextLineDistances[2]));
+			PedSimCityImage.distances.add(Float.parseFloat(nextLineDistances[2]));
 		}
 		readerDistances.close();
 	}
